test: Switch old Debian suites to archive
jessie-updates and jessie-backports are no longer on the mirrors
which is well documented: https://www.lucas-nussbaum.net/blog/?p=947

GCP image debian-8 has not had this done to it.
Do so in the default test scenario.

(This probably is a general purpose role to split off,
but doing it in the prepare play for now.)